+++ Kursovaya/src/main/java/com/travelAgency/Kursovaya/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<Admins,Long>{
+public interface UserRepository extends JpaRepository<Admins,Integer>{
     Admins findByUsername(String username);
 
 }
